+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 
 ### Added
+- contract structs & methods
+- property structs & methods
 
 ## [0.0.1 (2019-04-01)](https://github.com/hellocave/wazzup-sdk-go/releases/tag/0.0.1)